refactor(utils): move set_optimizer to model
